+++ packages/image_picker/image_picker_android/android/src/main/java/io/flutter/plugins/imagepicker/ImagePickerDelegate.java
@@ -304,7 +304,7 @@ public class ImagePickerDelegate
 
   private void launchPickMediaFromGalleryIntent(GeneralOptions generalOptions) {
     Intent pickMediaIntent;
-    if (generalOptions.getUsePhotoPicker()) {
+    if (ImagePickerUtils.effectiveUsePhotoPicker(generalOptions.getUsePhotoPicker())) {
       if (generalOptions.getAllowMultiple()) {
         int limit = ImagePickerUtils.getLimitFromOption(generalOptions);
 
@@ -353,7 +353,7 @@ public class ImagePickerDelegate
 
   private void launchPickVideoFromGalleryIntent(Boolean usePhotoPicker) {
     Intent pickVideoIntent;
-    if (usePhotoPicker) {
+    if (ImagePickerUtils.effectiveUsePhotoPicker(usePhotoPicker)) {
       pickVideoIntent =
           new ActivityResultContracts.PickVisualMedia()
               .createIntent(
@@ -462,7 +462,7 @@ public class ImagePickerDelegate
 
   private void launchPickImageFromGalleryIntent(Boolean usePhotoPicker) {
     Intent pickImageIntent;
-    if (usePhotoPicker) {
+    if (ImagePickerUtils.effectiveUsePhotoPicker(usePhotoPicker)) {
       pickImageIntent =
           new ActivityResultContracts.PickVisualMedia()
               .createIntent(
@@ -479,7 +479,7 @@ public class ImagePickerDelegate
 
   private void launchMultiPickImageFromGalleryIntent(Boolean usePhotoPicker, int limit) {
     Intent pickMultiImageIntent;
-    if (usePhotoPicker) {
+    if (ImagePickerUtils.effectiveUsePhotoPicker(usePhotoPicker)) {
       pickMultiImageIntent =
           new ActivityResultContracts.PickMultipleVisualMedia(limit)
               .createIntent(
@@ -514,7 +514,7 @@ public class ImagePickerDelegate
 
   private void launchMultiPickVideoFromGalleryIntent(Boolean usePhotoPicker, int limit) {
     Intent pickMultiVideoIntent;
-    if (usePhotoPicker) {
+    if (ImagePickerUtils.effectiveUsePhotoPicker(usePhotoPicker)) {
       pickMultiVideoIntent =
           new ActivityResultContracts.PickMultipleVisualMedia(limit)
               .createIntent(